pyproject.toml
[project]
name = "throatmicdata"
version = "0.1.0"
description = "Throat microphone dataset collection and management tools"
authors = [
    {name = "Paul Han", email = "pauljunsukhan@gmail.com"},
]
dependencies = [
    "numpy>=1.24.0",
    "pyaudio>=0.2.13",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "tqdm>=4.66.1",
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "huggingface-hub>=0.19.0",
    "pyyaml>=6.0.1",
    "datasets>=2.15.0",
    "librosa>=0.10.1",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend" 
